New Setup
New Script with imported modules.
Old Script renamed as archieve (delete after transformation is completed)
Integrated Setup of ParameterDomain
